requirements.txt
anyio==4.10.0
attrs==25.3.0
Automat==25.4.16
backports.tarfile==1.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
exceptiongroup==1.3.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
joblib==1.5.1
keyring==25.6.0
lightgbm==4.6.0
more-itertools==10.7.0
numpy==2.2.6
okx-sdk==5.4.1001
pandas==2.3.1
pycparser==2.22
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tomli==2.2.1
Twisted==25.5.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
xgboost==3.0.3
zipp==3.23.0
zope.interface==7.2
redis
pandas
pyarrow  # для Parquet
celery[redis]
redis